refactor: move playerlifestate to a different directory.
