+++ Modules/GameplaySession/src/Game.GameplaySession.Domain/Data/Entities/Session.cs
@@ -1,3 +1,4 @@
+using Game.GameplaySession.Domain.Data.Enums;
 using Game.GameplaySession.Domain.Data.ValueObjects;
 using Game.Shared.Core.Data;
