+++ Modules/GameplaySession/src/Game.GameplaySession.Domain/Data/Enums/PlayerLifeState.cs
@@ -1,7 +1,8 @@
-namespace Game.GameplaySession.Domain.Data.ValueObjects
+namespace Game.GameplaySession.Domain.Data.Enums
 {
     /// <summary>
     /// The life state of the player, either dead or alive.
+    /// TODO: This *should* map to a life state strategy class in the future, instead of being used directly.
     /// </summary>
     public enum PlayerLifeState
     {
